requirements.txt
Django==1.3
psycopg2==2.4.2
gunicorn==0.12.2
south==0.7.3
pil==1.1.7
simplejson==2.2.1
django-json-rpc==0.6.1